+++ CHANGELOG.md
@@ -20,3 +20,7 @@
 ## Jan 17, 2023
 
 - Backchannel logout (`backchannel_supported`) turned off for parent realm azure idp and standard realm azure idp in prod dev and prod test
+
+## June 19, 2023
+
+- Added new user_principal_name attribute for clients using Azure IDIR